requirements.txt
numpy
requests
websocket-client
tensorflow
pandas

